+++ dev/webdock/basebox/php-web-server-foundation/bin/entrypoint.bash
@@ -6,13 +6,16 @@ inject_environment_variables()
   envsubst $(printenv | cut -f1 -d'=' | sed 's/.*/\\\${&}/' | tr '\n' ',')
 }
 
-inject_environment_variables < /etc/nginx/nginx.template.conf > /etc/nginx/nginx.conf
-
-if [ "$XDEBUG_ENABLE" == 'on' ]
+if [ "$(id --user)" == '0' ]
 then
-  if php -r "exit(extension_loaded(\"xdebug\") ? 1 : 0);"
+  inject_environment_variables < /etc/nginx/nginx.template.conf > /etc/nginx/nginx.conf
+
+  if [ "$XDEBUG_ENABLE" == 'on' ]
   then
-    docker-php-ext-enable xdebug
+    if php -r "exit(extension_loaded(\"xdebug\") ? 1 : 0);"
+    then
+      docker-php-ext-enable xdebug
+    fi
   fi
 fi