+++ dev/webdock/basebox/web-assets-builder/hooks/test.sh
@@ -3,12 +3,9 @@ set -e
 . "$(dirname "$0")/_variables.sh"
 
 echo 'Running installation tests...'
-docker run --rm "${IMAGE}:${VERSION}" bash -c '
-  node --version \
-  && npm --version \
-  && git --version \
-  && sass --version \
-  && tsc --version \
-  && webpack --version \
-  && webpack-cli --version
+docker run --rm "${IMAGE}:${VERSION}" bash -ce '
+  sass --version
+  tsc --version
+  webpack --version
+  webpack-cli --version
 '