+++ dev/webdock/basebox/web-client-foundation/hooks/test.sh
@@ -3,8 +3,8 @@ set -e
 . "$(dirname "$0")/_variables.sh"
 
 echo 'Running installation tests...'
-docker run --rm "${IMAGE}:${VERSION}" bash -c '
-  nginx -v \
-  && curl --version \
-  && envsubst --version
+docker run --rm "${IMAGE}:${VERSION}" bash -ce '
+  nginx -v
+  curl --version
+  envsubst --version
 '