+++ dev/webdock/domain/php-web-server-tester/hooks/test.sh
@@ -5,8 +5,8 @@ set -e
 echo 'Running unit tests...'
 docker run \
   --rm \
-  --mount="type=volume,volume-driver=local,src=global_composer_cache,dst=/root/.composer/cache" \
-"${IMAGE}:${VERSION}" bash -c '
-  composer install --quiet \
-  && composer run-script unit-tests
+  --mount="type=volume,source=global_composer_cache,destination=/var/cache/composer" \
+"${IMAGE}:${VERSION}" bash -ce '
+  composer install --quiet
+  composer run-script unit-tests
 '