+++ dev/webdock/sandbox/standalone-firefox/hooks/test.sh
@@ -3,7 +3,7 @@ set -e
 . "$(dirname "$0")/_variables.sh"
 
 echo 'Running installation tests...'
-docker run --rm "${IMAGE}:${VERSION}" bash -c '
-  firefox --version \
-  && geckodriver --version
+docker run --rm "${IMAGE}:${VERSION}" bash -ce '
+  firefox --version
+  geckodriver --version
 '